Move get_children_by_key to parent module
